Make Settings page based
