+++ .eslintrc
@@ -1,6 +1,6 @@
 {
   "extends": ["next", "next/core-web-vitals"],
-  "rules": { 
-    "react/no-unescaped-entities": 0 
+  "rules": {
+    "react/no-unescaped-entities": 0
   }
 }
