+++ lib/MDXProvider.js
@@ -3,7 +3,14 @@
 import { MDXProvider } from "@mdx-js/react";
 
 // Design
-import { Heading, Text, Divider, UnorderedList, Code } from "@chakra-ui/react";
+import {
+  Heading,
+  Text,
+  Divider,
+  UnorderedList,
+  Code,
+  OrderedList,
+} from "@chakra-ui/react";
 
 const componentOverrides = {
   h1: (props) => <Heading size="xl" mb={5} {...props} />,
@@ -12,6 +19,7 @@ const componentOverrides = {
   p: (props) => <Text mb={2} {...props} />,
   hr: (props) => <Divider my={5} {...props} />,
   ul: (props) => <UnorderedList ps={5} {...props} />,
+  ol: (props) => <OrderedList ps={5} {...props} />,
   code: (props) => <Code {...props} />,
 };
