+++ public/markdown/docs/accessibility.mdx
@@ -1,6 +1,6 @@
 ---
 name: "Settings"
-date: "2022-05-19"
+date: "2022-06-03"
 ---
 
 ## Accessibility
@@ -13,12 +13,11 @@ We've implemented a handful of accessibility features, which are described below
 
 ### Font
 
-We use a custom font. To change the application font, go to Settings and select the "Change the Application Font" option. A modal will appear and offer you these options:
+The app uses a custom font. To enable or disable this behaviour, use Settings:
 
-- **Atkinson Hyperlegible:** A broad-based font made to put accessibility first. Developed by the Braille Institute of America and contributors.
-- **System Font:** This will tell the app to use the system font. Select this option if you have a custom font applied to the entire OS.
-
-To restore the default font, go to Settings, select "About the Osopcloud Application" and then "Reset Osopcloud".
+1. Go to Settings.
+2. Select "Advanced Settings".
+3. Enable (or disable) "Use System Font".
 
 ---
