+++ public/markdown/docs/getting-started.mdx
@@ -1,6 +1,6 @@
 ---
 name: "Getting Started"
-date: "2022-05-19"
+date: "2022-06-03"
 ---
 
 ## Getting Started
@@ -26,7 +26,7 @@ Home has two components. Both help you discover OSs:
 
 ### Operating System Pages
 
-Operating System Pages describe operating systems. There's be lot of information here:
+Operating System Pages describe operating systems. There's lot of information here:
 
 - **There's a description:** A detailed breakdown of the strengths and features of the OS.
 - **There's our metadata:** Our metadata gives you short and sharp insights across different aspects of the OS. Something special here is that lots of metadata is searchable with 8-way search.
