--- public/markdown/docs/introduction.mdx
@@ -1,14 +1,0 @@
----
-name: "Introduction"
-date: "2022-05-19"
----
-
-## Introduction
-
-Welcome to Osopcloud!
-
-It's a big and varied scene out there; with Linux, BSD, and more on the scene, each with their own strengths and features. So, we wanted to build a place where they can come together and discovering them is easy and modern.
-
-We want to have a set of tools here which both familiar and new users can use to culture build this community.
-
-Thanks team.
