+++ public/markdown/docs/settings.mdx
@@ -1,6 +1,6 @@
 ---
 name: "Settings"
-date: "2022-05-19"
+date: "2022-06-03"
 ---
 
 ## Settings
@@ -9,12 +9,24 @@ With Settings, you can customise and configure Osopcloud.
 
 Settings are stored on your device and remain the same every time you use Osopcloud.
 
-To go to Settings, select the gear icon button in the right of the navigation bar.
+To open Settings, select the gear icon button in the bottom right corner sidebar.
 
-There's a variety of settings available that change how Osopcloud looks. Some change looks, for example the session theme toggle. Others change how it works, for example, Focus Mode minimises notifications.
+---
+
+### Import or Export Settings
+
+If you need to take your Settings to another browser or device, use Osopcloud's built-in Settings import and export features.
+
+To export your settings, select "Manage Data or Reset" and then select "Export Settings Data to Clipboard". This will copy a JSON string to your clipboard which you can paste into a file or send to another device.
+
+To import settings, select "Manage Data or Reset" and then select "Import Settings from Clipboard". If prompted, select "Allow" or "Paste" to import the settings. This will overwrite your current settings.
+
+**Note:** When you use the Import Settings from Clipboard feature, you're full clipboard is revealed to the application. If there's no settings data in the clipboard, the import will fail.
+
+**Important:** It isn't possible to import settings on Mozilla Firefox.
 
 ---
 
 ### Reset Osopcloud
 
-To restore the default settings, select "About the Osopcloud Application" and then the "Reset Osopcloud" button.
+To restore the default settings, select "Manage Data or Reset" and then the "Reset Osopcloud" option.
