+++ public/markdown/docs/sharing.mdx
@@ -1,16 +1,14 @@
 ---
 name: "Share Points"
-date: "2022-05-19"
+date: "2022-06-03"
 ---
 
 ## Share Points
 
-Across Osopcloud, you see buttons with a "Share" label.
+Across Osopcloud, you see buttons with the share icon or the "Share" label.
 
 These are called Share Points.
 
 When you select a Share Point, the native share sheet appears. This allows you to share Osopcloud pages and content with the other apps on your system, or with your friends.
 
-Share Points is one of the many subtle culture building tools. By sharing open-source operating system culture with your apps and friends, you can help develop this community further.
-
-> **Note:** Share Points are hidden on browsers that don't support them, including Mozilla Firefox.
+Share Points are one of the many subtle culture building tools. By sharing open-source operating system culture with your apps and friends, you can help develop this community further.
