+++ public/sw.js
@@ -1,2 +1,2 @@
-if(!self.define){let e,s={};const a=(a,c)=>(a=new URL(a+".js",c).href,s[a]||new Promise((s=>{if("document"in self){const e=document.createElement("script");e.src=a,e.onload=s,document.head.appendChild(e)}else e=a,importScripts(a),s()})).then((()=>{let e=s[a];if(!e)throw new Error(`Module ${a} didn’t register its module`);return e})));self.define=(c,n)=>{const r=e||("document"in self?document.currentScript.src:"")||location.href;if(s[r])return;let i={};const d=e=>a(e,r),t={module:{uri:r},exports:i,require:d};s[r]=Promise.all(c.map((e=>t[e]||d(e)))).then((e=>(n(...e),i)))}}define(["./workbox-6316bd60"],(function(e){"use strict";importScripts("fallback-HA9m-E73CzrdYC9lLkccu.js"),self.addEventListener("message",(e=>{e.data&&"SKIP_WAITING"===e.data.type&&self.skipWaiting()})),e.clientsClaim(),e.precacheAndRoute([{url:"/_next/static/HA9m-E73CzrdYC9lLkccu/_buildManifest.js",revision:"6842c039f4aeac33de0e4472f639fa88"},{url:"/_next/static/HA9m-E73CzrdYC9lLkccu/_middlewareManifest.js",revision:"fb2823d66b3e778e04a3f681d0d2fb19"},{url:"/_next/static/HA9m-E73CzrdYC9lLkccu/_ssgManifest.js",revision:"b6652df95db52feb4daf4eca35380933"},{url:"/_next/static/chunks/280-d66d1b01ecde2017.js",revision:"d66d1b01ecde2017"},{url:"/_next/static/chunks/280-d66d1b01ecde2017.js.map",revision:"1a30ffb2103b66232fbc51955be04216"},{url:"/_next/static/chunks/868-08ee9e140a19ba96.js",revision:"08ee9e140a19ba96"},{url:"/_next/static/chunks/868-08ee9e140a19ba96.js.map",revision:"b7f0920c06ef670bb40730103ed92c82"},{url:"/_next/static/chunks/996-9c91d1eae613ddeb.js",revision:"9c91d1eae613ddeb"},{url:"/_next/static/chunks/996-9c91d1eae613ddeb.js.map",revision:"f8146419d29fa019023de28afea2bc50"},{url:"/_next/static/chunks/framework-da8ad0d3df154e7f.js",revision:"da8ad0d3df154e7f"},{url:"/_next/static/chunks/framework-da8ad0d3df154e7f.js.map",revision:"e43122222b79391baf565e0b2cea680d"},{url:"/_next/static/chunks/main-dda75576ff8dcbb9.js",revision:"dda75576ff8dcbb9"},{url:"/_next/static/chunks/main-dda75576ff8dcbb9.js.map",revision:"8ea4d1b06bf6684eb000fbfab9a361f1"},{url:"/_next/static/chunks/pages/404-ab4fc7984a47fc89.js",revision:"ab4fc7984a47fc89"},{url:"/_next/static/chunks/pages/404-ab4fc7984a47fc89.js.map",revision:"d9e8f0b91b552e0774b796e2f9182c5f"},{url:"/_next/static/chunks/pages/500-9bb490a10d186641.js",revision:"9bb490a10d186641"},{url:"/_next/static/chunks/pages/500-9bb490a10d186641.js.map",revision:"f3b7b2f244df3d2e7a1d9c8a0d3ba059"},{url:"/_next/static/chunks/pages/_app-ee71cf9f39fcebea.js",revision:"ee71cf9f39fcebea"},{url:"/_next/static/chunks/pages/_app-ee71cf9f39fcebea.js.map",revision:"c34c6c2b27efa9ea991c5053625fa720"},{url:"/_next/static/chunks/pages/_error-e2f2bfcd5131d4b7.js",revision:"e2f2bfcd5131d4b7"},{url:"/_next/static/chunks/pages/_error-e2f2bfcd5131d4b7.js.map",revision:"304eea13ca7ffb487de091a049524fe8"},{url:"/_next/static/chunks/pages/_offline-25f12faa815360c1.js",revision:"25f12faa815360c1"},{url:"/_next/static/chunks/pages/_offline-25f12faa815360c1.js.map",revision:"5922ec3cfd658921bb2f12cb60d51ab1"},{url:"/_next/static/chunks/pages/about/%5Bslug%5D-9555a2912bf68707.js",revision:"9555a2912bf68707"},{url:"/_next/static/chunks/pages/about/%5Bslug%5D-9555a2912bf68707.js.map",revision:"a9c1d62200d34b6a71c6dc0c4b9112d7"},{url:"/_next/static/chunks/pages/browse/%5Bslug%5D-d591cf368424a9de.js",revision:"d591cf368424a9de"},{url:"/_next/static/chunks/pages/browse/%5Bslug%5D-d591cf368424a9de.js.map",revision:"79d8fdd0c3c6350eb3620456b5fecbf7"},{url:"/_next/static/chunks/pages/commit-837588f7d4edcf5b.js",revision:"837588f7d4edcf5b"},{url:"/_next/static/chunks/pages/commit-837588f7d4edcf5b.js.map",revision:"5de01ad021835b7d3f17b1058fad3dcc"},{url:"/_next/static/chunks/pages/docs/%5Bslug%5D-913c1c67ec46da21.js",revision:"913c1c67ec46da21"},{url:"/_next/static/chunks/pages/docs/%5Bslug%5D-913c1c67ec46da21.js.map",revision:"d7e53c94b1ee8a3362e9008f5bd129ef"},{url:"/_next/static/chunks/pages/index-6b990afbd85c641e.js",revision:"6b990afbd85c641e"},{url:"/_next/static/chunks/pages/index-6b990afbd85c641e.js.map",revision:"7fa6d3f9bd8ecc9b5b293dfa80676f96"},{url:"/_next/static/chunks/pages/settings-fc32e95b7b7bb735.js",revision:"fc32e95b7b7bb735"},{url:"/_next/static/chunks/pages/settings-fc32e95b7b7bb735.js.map",revision:"ccd4248b1450e64d20da9382487d1ddd"},{url:"/_next/static/chunks/polyfills-5cd94c89d3acac5f.js",revision:"99442aec5788bccac9b2f0ead2afdd6b"},{url:"/_next/static/chunks/webpack-d13296a0cd7aff3f.js",revision:"d13296a0cd7aff3f"},{url:"/_next/static/chunks/webpack-d13296a0cd7aff3f.js.map",revision:"d2be422a478c50efc8d48e7ef2366ce7"},{url:"/_next/static/css/e6e39b401b232a51.css",revision:"e6e39b401b232a51"},{url:"/_next/static/css/e6e39b401b232a51.css.map",revision:"9060667fa8fe2e142ec54d3edfba8588"},{url:"/_next/static/media/public-sans-all-400-normal.ae05de44.woff",revision:"ae05de44"},{url:"/_next/static/media/public-sans-all-600-normal.07428ec5.woff",revision:"07428ec5"},{url:"/_next/static/media/public-sans-latin-400-normal.66c4eba4.woff2",revision:"66c4eba4"},{url:"/_next/static/media/public-sans-latin-600-normal.7e06741e.woff2",revision:"7e06741e"},{url:"/_next/static/media/public-sans-latin-ext-400-normal.fd422b1b.woff2",revision:"fd422b1b"},{url:"/_next/static/media/public-sans-latin-ext-600-normal.47b27dc8.woff2",revision:"47b27dc8"},{url:"/_offline",revision:"HA9m-E73CzrdYC9lLkccu"},{url:"/brand-assets/apple-icon-180.png",revision:"8780940006caca5b1e373f06e82fd1af"},{url:"/brand-assets/apple-splash-1125-2436.jpg",revision:"86fbd1cb121ceaf3e08bdcebba56c7be"},{url:"/brand-assets/apple-splash-1136-640.jpg",revision:"ede7a029c432b282e2cc14fe00239672"},{url:"/brand-assets/apple-splash-1170-2532.jpg",revision:"91d14a1e49019231cc28c998a7f698fd"},{url:"/brand-assets/apple-splash-1242-2208.jpg",revision:"31eeaa7d5bf26a403fb3a16774d6483e"},{url:"/brand-assets/apple-splash-1242-2688.jpg",revision:"7b495b042942bda0aa365bb7dbed8b3b"},{url:"/brand-assets/apple-splash-1284-2778.jpg",revision:"f87029e087d3020ffd0f35447c0fdf14"},{url:"/brand-assets/apple-splash-1334-750.jpg",revision:"165246c883305f86c45fd110818473bb"},{url:"/brand-assets/apple-splash-1536-2048.jpg",revision:"8d3dadbd41be1abac9bcbea2371dfd82"},{url:"/brand-assets/apple-splash-1620-2160.jpg",revision:"4ce8e458e1aa5fac5fb222fa7cb8cea9"},{url:"/brand-assets/apple-splash-1668-2224.jpg",revision:"5878b1a40b476a5364b562956d82f2c1"},{url:"/brand-assets/apple-splash-1668-2388.jpg",revision:"dfb78f53e7f2f292ec6f21639309663b"},{url:"/brand-assets/apple-splash-1792-828.jpg",revision:"c227af436352a6caaf93acb7f72c5c33"},{url:"/brand-assets/apple-splash-2048-1536.jpg",revision:"c4774c952cfb7d92873c477d3476de1a"},{url:"/brand-assets/apple-splash-2048-2732.jpg",revision:"8c8164473e2c039346bab0f04eb911e3"},{url:"/brand-assets/apple-splash-2160-1620.jpg",revision:"c36f5e613ea4fac80a8899354c85902e"},{url:"/brand-assets/apple-splash-2208-1242.jpg",revision:"275493f4cbab6255101666c01ee72b8c"},{url:"/brand-assets/apple-splash-2224-1668.jpg",revision:"1300f65f70553e9392a80973ed272474"},{url:"/brand-assets/apple-splash-2388-1668.jpg",revision:"f2090ccbd3b7562ac39439f723c7136f"},{url:"/brand-assets/apple-splash-2436-1125.jpg",revision:"d2f5d6ec44eb0e0c8c4e0cedd690bd05"},{url:"/brand-assets/apple-splash-2532-1170.jpg",revision:"efac034a67fbc6cc14f8f35a12a3456a"},{url:"/brand-assets/apple-splash-2688-1242.jpg",revision:"7c90bdff925d7fcf9ab580a17d295d6a"},{url:"/brand-assets/apple-splash-2732-2048.jpg",revision:"516acb6571386fdf76de28686f937e87"},{url:"/brand-assets/apple-splash-2778-1284.jpg",revision:"efe493e5a9fc277a380c209b9e97d1c9"},{url:"/brand-assets/apple-splash-640-1136.jpg",revision:"069cd3496edcac61154a939ae02bb835"},{url:"/brand-assets/apple-splash-750-1334.jpg",revision:"e106f8b20677b8d7df0f0e2dc61824c3"},{url:"/brand-assets/apple-splash-828-1792.jpg",revision:"296bdce7a826fa315ad7e3cb115ee0c8"},{url:"/brand-assets/favicon-196.png",revision:"1405dff5bbadc38724b93ad4c0a8f23f"},{url:"/brand-assets/manifest-icon-192.png",revision:"e4312fb3f6e55c8b7e1070bdc2522f52"},{url:"/brand-assets/manifest-icon-512.png",revision:"40a35280399add43df9c56753b332bbe"},{url:"/manifest.json",revision:"53149ae646cff0058e467800c1a5be88"},{url:"/markdown/about/privacy.mdx",revision:"e8add1238d2aa01184ad7dfdb52cef08"},{url:"/markdown/about/terms.mdx",revision:"7df0a04cf2ef2a9b794bac9a3336d095"},{url:"/markdown/browse/demo.mdx",revision:"7fc8f08ea2c2a279ac1dd63e6f09a3c7"},{url:"/markdown/docs/accessibility.mdx",revision:"724a2f0290f63502e0ceaf655780c6cf"},{url:"/markdown/docs/errors.mdx",revision:"f3dfb09ebb36f71a4b7c9844456d0bbe"},{url:"/markdown/docs/getting-started.mdx",revision:"a509320283f7d88387afc658990f81b2"},{url:"/markdown/docs/introduction.mdx",revision:"59f32c7b17533d9eaa4473fa48a5cac5"},{url:"/markdown/docs/keyboard-shortcuts.mdx",revision:"a7a271c9daf699e16be40c7812fae65f"},{url:"/markdown/docs/settings.mdx",revision:"23ca6ddf6cf2dc7bec50df099f754f4d"},{url:"/markdown/docs/sharing.mdx",revision:"599366d8e0d7e56742a1859960c78e41"}],{ignoreURLParametersMatching:[]}),e.cleanupOutdatedCaches(),e.registerRoute("/",new e.NetworkFirst({cacheName:"start-url",plugins:[{cacheWillUpdate:async({request:e,response:s,event:a,state:c})=>s&&"opaqueredirect"===s.type?new Response(s.body,{status:200,statusText:"OK",headers:s.headers}):s},{handlerDidError:async({request:e})=>self.fallback(e)}]}),"GET"),e.registerRoute(/^https:\/\/fonts\.(?:gstatic)\.com\/.*/i,new e.CacheFirst({cacheName:"google-fonts-webfonts",plugins:[new e.ExpirationPlugin({maxEntries:4,maxAgeSeconds:31536e3}),{handlerDidError:async({request:e})=>self.fallback(e)}]}),"GET"),e.registerRoute(/^https:\/\/fonts\.(?:googleapis)\.com\/.*/i,new e.StaleWhileRevalidate({cacheName:"google-fonts-stylesheets",plugins:[new e.ExpirationPlugin({maxEntries:4,maxAgeSeconds:604800}),{handlerDidError:async({request:e})=>self.fallback(e)}]}),"GET"),e.registerRoute(/\.(?:eot|otf|ttc|ttf|woff|woff2|font.css)$/i,new e.StaleWhileRevalidate({cacheName:"static-font-assets",plugins:[new e.ExpirationPlugin({maxEntries:4,maxAgeSeconds:604800}),{handlerDidError:async({request:e})=>self.fallback(e)}]}),"GET"),e.registerRoute(/\.(?:jpg|jpeg|gif|png|svg|ico|webp)$/i,new e.StaleWhileRevalidate({cacheName:"static-image-assets",plugins:[new e.ExpirationPlugin({maxEntries:64,maxAgeSeconds:86400}),{handlerDidError:async({request:e})=>self.fallback(e)}]}),"GET"),e.registerRoute(/\/_next\/image\?url=.+$/i,new e.StaleWhileRevalidate({cacheName:"next-image",plugins:[new e.ExpirationPlugin({maxEntries:64,maxAgeSeconds:86400}),{handlerDidError:async({request:e})=>self.fallback(e)}]}),"GET"),e.registerRoute(/\.(?:mp3|wav|ogg)$/i,new e.CacheFirst({cacheName:"static-audio-assets",plugins:[new e.RangeRequestsPlugin,new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400}),{handlerDidError:async({request:e})=>self.fallback(e)}]}),"GET"),e.registerRoute(/\.(?:mp4)$/i,new e.CacheFirst({cacheName:"static-video-assets",plugins:[new e.RangeRequestsPlugin,new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400}),{handlerDidError:async({request:e})=>self.fallback(e)}]}),"GET"),e.registerRoute(/\.(?:js)$/i,new e.StaleWhileRevalidate({cacheName:"static-js-assets",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400}),{handlerDidError:async({request:e})=>self.fallback(e)}]}),"GET"),e.registerRoute(/\.(?:css|less)$/i,new e.StaleWhileRevalidate({cacheName:"static-style-assets",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400}),{handlerDidError:async({request:e})=>self.fallback(e)}]}),"GET"),e.registerRoute(/\/_next\/data\/.+\/.+\.json$/i,new e.StaleWhileRevalidate({cacheName:"next-data",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400}),{handlerDidError:async({request:e})=>self.fallback(e)}]}),"GET"),e.registerRoute(/\.(?:json|xml|csv)$/i,new e.NetworkFirst({cacheName:"static-data-assets",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400}),{handlerDidError:async({request:e})=>self.fallback(e)}]}),"GET"),e.registerRoute((({url:e})=>{if(!(self.origin===e.origin))return!1;const s=e.pathname;return!s.startsWith("/api/auth/")&&!!s.startsWith("/api/")}),new e.NetworkFirst({cacheName:"apis",networkTimeoutSeconds:10,plugins:[new e.ExpirationPlugin({maxEntries:16,maxAgeSeconds:86400}),{handlerDidError:async({request:e})=>self.fallback(e)}]}),"GET"),e.registerRoute((({url:e})=>{if(!(self.origin===e.origin))return!1;return!e.pathname.startsWith("/api/")}),new e.NetworkFirst({cacheName:"others",networkTimeoutSeconds:10,plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400}),{handlerDidError:async({request:e})=>self.fallback(e)}]}),"GET"),e.registerRoute((({url:e})=>!(self.origin===e.origin)),new e.NetworkFirst({cacheName:"cross-origin",networkTimeoutSeconds:10,plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:3600}),{handlerDidError:async({request:e})=>self.fallback(e)}]}),"GET")}));
+if(!self.define){let e,s={};const a=(a,n)=>(a=new URL(a+".js",n).href,s[a]||new Promise((s=>{if("document"in self){const e=document.createElement("script");e.src=a,e.onload=s,document.head.appendChild(e)}else e=a,importScripts(a),s()})).then((()=>{let e=s[a];if(!e)throw new Error(`Module ${a} didn’t register its module`);return e})));self.define=(n,c)=>{const i=e||("document"in self?document.currentScript.src:"")||location.href;if(s[i])return;let r={};const d=e=>a(e,i),t={module:{uri:i},exports:r,require:d};s[i]=Promise.all(n.map((e=>t[e]||d(e)))).then((e=>(c(...e),r)))}}define(["./workbox-6316bd60"],(function(e){"use strict";importScripts("fallback-S7KBGSIHkQu74l79W0Yfk.js"),self.addEventListener("message",(e=>{e.data&&"SKIP_WAITING"===e.data.type&&self.skipWaiting()})),e.clientsClaim(),e.precacheAndRoute([{url:"/_next/static/S7KBGSIHkQu74l79W0Yfk/_buildManifest.js",revision:"bc94a8d33d2e5e3446f009a9b9e0f814"},{url:"/_next/static/S7KBGSIHkQu74l79W0Yfk/_middlewareManifest.js",revision:"fb2823d66b3e778e04a3f681d0d2fb19"},{url:"/_next/static/S7KBGSIHkQu74l79W0Yfk/_ssgManifest.js",revision:"b6652df95db52feb4daf4eca35380933"},{url:"/_next/static/chunks/381-0afc9b68ef24053d.js",revision:"0afc9b68ef24053d"},{url:"/_next/static/chunks/381-0afc9b68ef24053d.js.map",revision:"00d61e811953426571fb7c75c391bd0d"},{url:"/_next/static/chunks/808-363871de133561f7.js",revision:"363871de133561f7"},{url:"/_next/static/chunks/808-363871de133561f7.js.map",revision:"8009fd592f676e42622800e220da34c2"},{url:"/_next/static/chunks/845-d47e4d637588e7b1.js",revision:"d47e4d637588e7b1"},{url:"/_next/static/chunks/845-d47e4d637588e7b1.js.map",revision:"9ce9189e6c2f1b37ebed5429bf7cc75b"},{url:"/_next/static/chunks/868-e4df72b122f0096f.js",revision:"e4df72b122f0096f"},{url:"/_next/static/chunks/868-e4df72b122f0096f.js.map",revision:"156ad1d4e4b41c939979030d86bfc8a6"},{url:"/_next/static/chunks/996-9c91d1eae613ddeb.js",revision:"9c91d1eae613ddeb"},{url:"/_next/static/chunks/996-9c91d1eae613ddeb.js.map",revision:"f8146419d29fa019023de28afea2bc50"},{url:"/_next/static/chunks/framework-da8ad0d3df154e7f.js",revision:"da8ad0d3df154e7f"},{url:"/_next/static/chunks/framework-da8ad0d3df154e7f.js.map",revision:"e43122222b79391baf565e0b2cea680d"},{url:"/_next/static/chunks/main-dda75576ff8dcbb9.js",revision:"dda75576ff8dcbb9"},{url:"/_next/static/chunks/main-dda75576ff8dcbb9.js.map",revision:"8ea4d1b06bf6684eb000fbfab9a361f1"},{url:"/_next/static/chunks/pages/404-2826225aa793b388.js",revision:"2826225aa793b388"},{url:"/_next/static/chunks/pages/404-2826225aa793b388.js.map",revision:"89a1b24036084253cbded1b09623a77f"},{url:"/_next/static/chunks/pages/500-2b436cdf2fad56ee.js",revision:"2b436cdf2fad56ee"},{url:"/_next/static/chunks/pages/500-2b436cdf2fad56ee.js.map",revision:"098f073cfb838553ba40260e6e18ab8e"},{url:"/_next/static/chunks/pages/_app-75d97c8dd2183bd3.js",revision:"75d97c8dd2183bd3"},{url:"/_next/static/chunks/pages/_app-75d97c8dd2183bd3.js.map",revision:"b598b49de576a7ff373bdf829118e1d7"},{url:"/_next/static/chunks/pages/_error-e2f2bfcd5131d4b7.js",revision:"e2f2bfcd5131d4b7"},{url:"/_next/static/chunks/pages/_error-e2f2bfcd5131d4b7.js.map",revision:"304eea13ca7ffb487de091a049524fe8"},{url:"/_next/static/chunks/pages/_offline-7cf650bfb704d70a.js",revision:"7cf650bfb704d70a"},{url:"/_next/static/chunks/pages/_offline-7cf650bfb704d70a.js.map",revision:"8342aa437f344f1603ed5a08474d7477"},{url:"/_next/static/chunks/pages/about/%5Bslug%5D-9555a2912bf68707.js",revision:"9555a2912bf68707"},{url:"/_next/static/chunks/pages/about/%5Bslug%5D-9555a2912bf68707.js.map",revision:"a9c1d62200d34b6a71c6dc0c4b9112d7"},{url:"/_next/static/chunks/pages/browse/%5Bslug%5D-a9bf1d1df98c1756.js",revision:"a9bf1d1df98c1756"},{url:"/_next/static/chunks/pages/browse/%5Bslug%5D-a9bf1d1df98c1756.js.map",revision:"bfe1cc49c258386112d543991599f5cd"},{url:"/_next/static/chunks/pages/commit-837588f7d4edcf5b.js",revision:"837588f7d4edcf5b"},{url:"/_next/static/chunks/pages/commit-837588f7d4edcf5b.js.map",revision:"5de01ad021835b7d3f17b1058fad3dcc"},{url:"/_next/static/chunks/pages/docs/%5Bslug%5D-5e7b29b098ea7b75.js",revision:"5e7b29b098ea7b75"},{url:"/_next/static/chunks/pages/docs/%5Bslug%5D-5e7b29b098ea7b75.js.map",revision:"b66115d62ab4113bc0c632a2631d03d5"},{url:"/_next/static/chunks/pages/index-4870817b2d3bc1e6.js",revision:"4870817b2d3bc1e6"},{url:"/_next/static/chunks/pages/index-4870817b2d3bc1e6.js.map",revision:"be0626139aca816a87aa7ecca341d105"},{url:"/_next/static/chunks/pages/settings/advanced-f3a8b37586d62f00.js",revision:"f3a8b37586d62f00"},{url:"/_next/static/chunks/pages/settings/advanced-f3a8b37586d62f00.js.map",revision:"e9f6e906da834882b291877a1abf3ae3"},{url:"/_next/static/chunks/pages/settings/general-b1bc341549b6489f.js",revision:"b1bc341549b6489f"},{url:"/_next/static/chunks/pages/settings/general-b1bc341549b6489f.js.map",revision:"ce011905c9233870fe81fa8a3b6409b3"},{url:"/_next/static/chunks/pages/settings/manage-data-02966bbf64c5653d.js",revision:"02966bbf64c5653d"},{url:"/_next/static/chunks/pages/settings/manage-data-02966bbf64c5653d.js.map",revision:"afd22d915325fa9b96ef0c9be2e4dba2"},{url:"/_next/static/chunks/polyfills-5cd94c89d3acac5f.js",revision:"99442aec5788bccac9b2f0ead2afdd6b"},{url:"/_next/static/chunks/webpack-4e27c9d56dcd8b5a.js",revision:"4e27c9d56dcd8b5a"},{url:"/_next/static/chunks/webpack-4e27c9d56dcd8b5a.js.map",revision:"5e912d932217baf1742147489d1bb0a0"},{url:"/_next/static/css/e6e39b401b232a51.css",revision:"e6e39b401b232a51"},{url:"/_next/static/css/e6e39b401b232a51.css.map",revision:"9060667fa8fe2e142ec54d3edfba8588"},{url:"/_next/static/media/public-sans-all-400-normal.ae05de44.woff",revision:"ae05de44"},{url:"/_next/static/media/public-sans-all-600-normal.07428ec5.woff",revision:"07428ec5"},{url:"/_next/static/media/public-sans-latin-400-normal.66c4eba4.woff2",revision:"66c4eba4"},{url:"/_next/static/media/public-sans-latin-600-normal.7e06741e.woff2",revision:"7e06741e"},{url:"/_next/static/media/public-sans-latin-ext-400-normal.fd422b1b.woff2",revision:"fd422b1b"},{url:"/_next/static/media/public-sans-latin-ext-600-normal.47b27dc8.woff2",revision:"47b27dc8"},{url:"/_offline",revision:"S7KBGSIHkQu74l79W0Yfk"},{url:"/brand-assets/apple-icon-180.png",revision:"8780940006caca5b1e373f06e82fd1af"},{url:"/brand-assets/apple-splash-1125-2436.jpg",revision:"86fbd1cb121ceaf3e08bdcebba56c7be"},{url:"/brand-assets/apple-splash-1136-640.jpg",revision:"ede7a029c432b282e2cc14fe00239672"},{url:"/brand-assets/apple-splash-1170-2532.jpg",revision:"91d14a1e49019231cc28c998a7f698fd"},{url:"/brand-assets/apple-splash-1242-2208.jpg",revision:"31eeaa7d5bf26a403fb3a16774d6483e"},{url:"/brand-assets/apple-splash-1242-2688.jpg",revision:"7b495b042942bda0aa365bb7dbed8b3b"},{url:"/brand-assets/apple-splash-1284-2778.jpg",revision:"f87029e087d3020ffd0f35447c0fdf14"},{url:"/brand-assets/apple-splash-1334-750.jpg",revision:"165246c883305f86c45fd110818473bb"},{url:"/brand-assets/apple-splash-1536-2048.jpg",revision:"8d3dadbd41be1abac9bcbea2371dfd82"},{url:"/brand-assets/apple-splash-1620-2160.jpg",revision:"4ce8e458e1aa5fac5fb222fa7cb8cea9"},{url:"/brand-assets/apple-splash-1668-2224.jpg",revision:"5878b1a40b476a5364b562956d82f2c1"},{url:"/brand-assets/apple-splash-1668-2388.jpg",revision:"dfb78f53e7f2f292ec6f21639309663b"},{url:"/brand-assets/apple-splash-1792-828.jpg",revision:"c227af436352a6caaf93acb7f72c5c33"},{url:"/brand-assets/apple-splash-2048-1536.jpg",revision:"c4774c952cfb7d92873c477d3476de1a"},{url:"/brand-assets/apple-splash-2048-2732.jpg",revision:"8c8164473e2c039346bab0f04eb911e3"},{url:"/brand-assets/apple-splash-2160-1620.jpg",revision:"c36f5e613ea4fac80a8899354c85902e"},{url:"/brand-assets/apple-splash-2208-1242.jpg",revision:"275493f4cbab6255101666c01ee72b8c"},{url:"/brand-assets/apple-splash-2224-1668.jpg",revision:"1300f65f70553e9392a80973ed272474"},{url:"/brand-assets/apple-splash-2388-1668.jpg",revision:"f2090ccbd3b7562ac39439f723c7136f"},{url:"/brand-assets/apple-splash-2436-1125.jpg",revision:"d2f5d6ec44eb0e0c8c4e0cedd690bd05"},{url:"/brand-assets/apple-splash-2532-1170.jpg",revision:"efac034a67fbc6cc14f8f35a12a3456a"},{url:"/brand-assets/apple-splash-2688-1242.jpg",revision:"7c90bdff925d7fcf9ab580a17d295d6a"},{url:"/brand-assets/apple-splash-2732-2048.jpg",revision:"516acb6571386fdf76de28686f937e87"},{url:"/brand-assets/apple-splash-2778-1284.jpg",revision:"efe493e5a9fc277a380c209b9e97d1c9"},{url:"/brand-assets/apple-splash-640-1136.jpg",revision:"069cd3496edcac61154a939ae02bb835"},{url:"/brand-assets/apple-splash-750-1334.jpg",revision:"e106f8b20677b8d7df0f0e2dc61824c3"},{url:"/brand-assets/apple-splash-828-1792.jpg",revision:"296bdce7a826fa315ad7e3cb115ee0c8"},{url:"/brand-assets/favicon-196.png",revision:"1405dff5bbadc38724b93ad4c0a8f23f"},{url:"/brand-assets/manifest-icon-192.png",revision:"e4312fb3f6e55c8b7e1070bdc2522f52"},{url:"/brand-assets/manifest-icon-512.png",revision:"40a35280399add43df9c56753b332bbe"},{url:"/manifest.json",revision:"53149ae646cff0058e467800c1a5be88"},{url:"/markdown/about/privacy.mdx",revision:"e8add1238d2aa01184ad7dfdb52cef08"},{url:"/markdown/about/terms.mdx",revision:"7df0a04cf2ef2a9b794bac9a3336d095"},{url:"/markdown/browse/demo.mdx",revision:"7fc8f08ea2c2a279ac1dd63e6f09a3c7"},{url:"/markdown/docs/accessibility.mdx",revision:"09e7b182ca58740dd17ee9c2789c5b04"},{url:"/markdown/docs/errors.mdx",revision:"f3dfb09ebb36f71a4b7c9844456d0bbe"},{url:"/markdown/docs/getting-started.mdx",revision:"b8062fe743cdc7d801d09d5615b51dc0"},{url:"/markdown/docs/keyboard-shortcuts.mdx",revision:"a7a271c9daf699e16be40c7812fae65f"},{url:"/markdown/docs/settings.mdx",revision:"9552b6dc10c6383ebdfeae9e24f994ca"},{url:"/markdown/docs/sharing.mdx",revision:"dd0a2ca870d54f016f235756fddc9dc6"}],{ignoreURLParametersMatching:[]}),e.cleanupOutdatedCaches(),e.registerRoute("/",new e.NetworkFirst({cacheName:"start-url",plugins:[{cacheWillUpdate:async({request:e,response:s,event:a,state:n})=>s&&"opaqueredirect"===s.type?new Response(s.body,{status:200,statusText:"OK",headers:s.headers}):s},{handlerDidError:async({request:e})=>self.fallback(e)}]}),"GET"),e.registerRoute(/^https:\/\/fonts\.(?:gstatic)\.com\/.*/i,new e.CacheFirst({cacheName:"google-fonts-webfonts",plugins:[new e.ExpirationPlugin({maxEntries:4,maxAgeSeconds:31536e3}),{handlerDidError:async({request:e})=>self.fallback(e)}]}),"GET"),e.registerRoute(/^https:\/\/fonts\.(?:googleapis)\.com\/.*/i,new e.StaleWhileRevalidate({cacheName:"google-fonts-stylesheets",plugins:[new e.ExpirationPlugin({maxEntries:4,maxAgeSeconds:604800}),{handlerDidError:async({request:e})=>self.fallback(e)}]}),"GET"),e.registerRoute(/\.(?:eot|otf|ttc|ttf|woff|woff2|font.css)$/i,new e.StaleWhileRevalidate({cacheName:"static-font-assets",plugins:[new e.ExpirationPlugin({maxEntries:4,maxAgeSeconds:604800}),{handlerDidError:async({request:e})=>self.fallback(e)}]}),"GET"),e.registerRoute(/\.(?:jpg|jpeg|gif|png|svg|ico|webp)$/i,new e.StaleWhileRevalidate({cacheName:"static-image-assets",plugins:[new e.ExpirationPlugin({maxEntries:64,maxAgeSeconds:86400}),{handlerDidError:async({request:e})=>self.fallback(e)}]}),"GET"),e.registerRoute(/\/_next\/image\?url=.+$/i,new e.StaleWhileRevalidate({cacheName:"next-image",plugins:[new e.ExpirationPlugin({maxEntries:64,maxAgeSeconds:86400}),{handlerDidError:async({request:e})=>self.fallback(e)}]}),"GET"),e.registerRoute(/\.(?:mp3|wav|ogg)$/i,new e.CacheFirst({cacheName:"static-audio-assets",plugins:[new e.RangeRequestsPlugin,new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400}),{handlerDidError:async({request:e})=>self.fallback(e)}]}),"GET"),e.registerRoute(/\.(?:mp4)$/i,new e.CacheFirst({cacheName:"static-video-assets",plugins:[new e.RangeRequestsPlugin,new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400}),{handlerDidError:async({request:e})=>self.fallback(e)}]}),"GET"),e.registerRoute(/\.(?:js)$/i,new e.StaleWhileRevalidate({cacheName:"static-js-assets",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400}),{handlerDidError:async({request:e})=>self.fallback(e)}]}),"GET"),e.registerRoute(/\.(?:css|less)$/i,new e.StaleWhileRevalidate({cacheName:"static-style-assets",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400}),{handlerDidError:async({request:e})=>self.fallback(e)}]}),"GET"),e.registerRoute(/\/_next\/data\/.+\/.+\.json$/i,new e.StaleWhileRevalidate({cacheName:"next-data",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400}),{handlerDidError:async({request:e})=>self.fallback(e)}]}),"GET"),e.registerRoute(/\.(?:json|xml|csv)$/i,new e.NetworkFirst({cacheName:"static-data-assets",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400}),{handlerDidError:async({request:e})=>self.fallback(e)}]}),"GET"),e.registerRoute((({url:e})=>{if(!(self.origin===e.origin))return!1;const s=e.pathname;return!s.startsWith("/api/auth/")&&!!s.startsWith("/api/")}),new e.NetworkFirst({cacheName:"apis",networkTimeoutSeconds:10,plugins:[new e.ExpirationPlugin({maxEntries:16,maxAgeSeconds:86400}),{handlerDidError:async({request:e})=>self.fallback(e)}]}),"GET"),e.registerRoute((({url:e})=>{if(!(self.origin===e.origin))return!1;return!e.pathname.startsWith("/api/")}),new e.NetworkFirst({cacheName:"others",networkTimeoutSeconds:10,plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400}),{handlerDidError:async({request:e})=>self.fallback(e)}]}),"GET"),e.registerRoute((({url:e})=>!(self.origin===e.origin)),new e.NetworkFirst({cacheName:"cross-origin",networkTimeoutSeconds:10,plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:3600}),{handlerDidError:async({request:e})=>self.fallback(e)}]}),"GET")}));
 //# sourceMappingURL=sw.js.map
